Refactor operations over insights settings (settings cascade)
