+++ client/web/src/insights/pages/creation/lang-stats/LangStatsInsightCreationPage.tsx
@@ -13,7 +13,7 @@ import { Page } from '../../../../components/Page'
 import { PageTitle } from '../../../../components/PageTitle'
 import { FORM_ERROR, FormChangeEvent } from '../../../components/form/hooks/useForm'
 import { InsightsApiContext } from '../../../core/backend/api-provider'
-import { addInsightToCascadeSetting } from '../../../core/jsonc-operation'
+import { addInsightToSettings } from '../../../core/settings-action/insights'
 
 import {
     LangStatsInsightCreationContent,
@@ -67,7 +67,7 @@ export const LangStatsInsightCreationPage: React.FunctionComponent<LangStatsInsi
                 const settings = await getSubjectSettings(subjectID).toPromise()
 
                 const insight = getSanitizedLangStatsInsight(values)
-                const editedSettings = addInsightToCascadeSetting(settings.contents, insight)
+                const editedSettings = addInsightToSettings(settings.contents, insight)
 
                 await updateSubjectSettings(platformContext, subjectID, editedSettings).toPromise()
